move styles out of assets dir, update imports
