+++ src/components/BottomBar/BottomBar.tsx
@@ -4,7 +4,7 @@ import { useUserContext } from '../../usercontext';
 import SplitButton from '../SplitButton/SplitButton';
 import DefectiveDialog from '../DefectiveDialog/DefectiveDialog';
 import { BottombarProps } from '../../types';
-import { BottomBarStyles as styles } from '../../assets/styles';
+import { BottomBarStyles as styles } from '../../styles';
 import { Grid, Box, Paper, Button, CssBaseline } from '@mui/material';
 import KeyboardArrowLeftIcon from '@mui/icons-material/KeyboardArrowLeft';
 import KeyboardArrowRightIcon from '@mui/icons-material/KeyboardArrowRight';
