+++ src/components/ProjectsListTable/ProjectsListTable.tsx
@@ -2,7 +2,7 @@ import { useNavigate } from "react-router-dom";
 import { Table, TableBody, TableCell, TableContainer, TableHead, TableRow, Paper } from '@mui/material';
 import { formatDate } from '../../util';
 import { ProjectData } from "../../types";
-import { styles } from "../../assets/styles";
+import { styles } from "../../styles";
 
 interface ProjectsListTableProps {
   projects: ProjectData[];
